third_party/WebKit/LayoutTests/fast/js/unscopeable-expected.txt

Test [Unscopeable] attributes and methods.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "unscopeableAttribute" in internals is true
PASS "unscopeableMethod" in internals is true
PASS internals.unscopeableAttribute is "unscopeableAttribute"
PASS internals.unscopeableMethod() is "unscopeableMethod"
PASS unscopeableAttribute is "variable"
PASS unscopeableMethod is "variable"
PASS successfullyParsed is true

TEST COMPLETE
